+++ lib/plex_annotations/plex_annotations.dart
@@ -1,0 +1,6 @@
+
+class PlexAnnotationModel {
+  const PlexAnnotationModel();
+}
+
+const plexAnnotationModel = PlexAnnotationModel();