+++ lib/plex_annotations/plex_builders/plex_annotation_builders.dart
@@ -1,0 +1,11 @@
+import 'package:build/build.dart';
+import 'package:plex/plex_annotations/plex_generators/plex_model_generators.dart';
+import 'package:source_gen/source_gen.dart';
+
+Builder generateJsonClass(BuilderOptions options) {
+  return PartBuilder(
+    [PlexModelGenerator()],
+    '.plex.dart',
+    allowSyntaxErrors: true,
+  );
+}