+++ lib/plex_widgets/plex_data_table_paginated.dart
@@ -8,6 +8,10 @@ import 'package:plex/plex_widgets/plex_data_table.dart';
 import 'package:plex/plex_widgets/plex_input_widget.dart';
 import 'package:plex/plex_widgets/plex_shimmer.dart';
 
+
+/// [PlexDataTable] class is deprecated and should no longer be used.
+/// Please use [PlexAdvanceDataTable] instead
+@Deprecated("Use PlexAdvanceDataTable instead")
 class PlexDataTableWithPages extends StatefulWidget {
   PlexDataTableWithPages({
     super.key,